requirements.txt
chalice
line-bot-sdk
requests
